Changed back to int from long.

+++ yukon-client/common/src/com/cannontech/database/incrementer/KeyedIncrementer.java
@@ -1,5 +1,5 @@
 package com.cannontech.database.incrementer;
 
 public interface KeyedIncrementer {
-    public long getNextValue(String key);
+    public int getNextValue(String key);
 }
